Revert "build: rerun build.rs if the downloaded pre-built archive file has changed"
This reverts commit fc802905ce33aea702d5ea5cd6b07510e1f96af4, which
causes redundant recompilation. Cargo compares the modification time
(mtime) of downloaded_file against that of `output` in order to
determine if downloaded_file has changed. The mtime of `output` is set
to the start time of the build script:

* https://github.com/rust-lang/cargo/blob/247b22f225af3fbdb82e53e2f1ab81efd3cab050/src/cargo/core/compiler/custom_build.rs#L386
* https://github.com/rust-lang/cargo/blob/247b22f225af3fbdb82e53e2f1ab81efd3cab050/src/cargo/core/compiler/custom_build.rs#L432

Therefore, the mtime of `output` is earlier than that of downloaded_file
when downloaded_file is newly downloaded during the build script.